Twilio Removed

+++ backend/app/Http/Controllers/UserController.php
@@ -7,8 +7,11 @@ use Laravel\Sanctum\HasApiTokens;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\ValidationException;
+use App\Mail\SendCredentials;
+use Illuminate\Support\Facades\Mail;
 use App\Models\Student;
 use App\Models\User;
+use App\Models\Order;
 use App\Models\Menu;
 use App\Models\Permission;
 use App\Models\Role;
@@ -29,15 +32,40 @@ class UserController extends Controller
         return response()->json(Auth::user()->only(['firstname', 'lastname', 'picture', 'role']));
     }
 
+    public function setupAccount(Request $request) {
+        $user = User::with('role')->where('id', $request->id)->first();
+        return response()->json($user);
+    }
+
+    public function saveSetup(Request $request) {
+        try {
+            $user = User::find($request->id);
+            $validate = $request->validate([
+                'password' => 'required'
+            ]);
+
+            $user->password = $request->password;
+            $user->update();
+
+            return response()->json([
+                'message' => 'Saved successfully'
+            ]);
+        }catch(ValidationException $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ]);
+        }
+    }
+
     public function showUsers() {
         try {
             $currentUser = User::with('role')->find(Auth::id());
 
             if($currentUser->role->name =='Admin'){
-                $users = User::with('role', 'permissions')
+                $users = User::with('role')
                 ->where('role_id', '!=', 1)->get();
             }else{
-                $users = User::with('role', 'permissions')
+                $users = User::with('role')
                 ->where('role_id', '!=', 1)
                 ->where('role_id', '!=', $currentUser->role_id)
                 ->get();
@@ -49,7 +77,6 @@ class UserController extends Controller
             return response()->json([
                 'users' => $users,
                 'currentUser' => $currentUser->role->name,
-                'permissions' => $permissions,
                 'roles' => $roles
             ], 200);
 
@@ -75,6 +102,11 @@ class UserController extends Controller
             $credentials = User::where('email', $request->email)->first();
             if($credentials && Hash::check($request->password, $credentials->password)) {
                 $token = $credentials->createToken('personal-token')->plainTextToken;
+                if($credentials->status == null) {
+                    $credentials->status = 'active';
+                }
+                $credentials->save();
+
                 return response()->json([
                     'token' => $token,
                     'data' => $credentials
@@ -93,6 +125,14 @@ class UserController extends Controller
 
     function logout(Request $request){
         $user = Auth::user();
+        if($user->status == 'Assigned') {
+            return response()->json([
+                'message' => 'You cannot log out while a delivery is in progress. Please complete or ask for reassignment of the delivery first.'
+            ]);
+        }
+
+        $user->status = null;
+        $user->save();
         $user->tokens()->delete();
         return response()->json([
             'success' => 'Logout successfully'
@@ -118,16 +158,16 @@ class UserController extends Controller
                 'role_id' => 'required',
                 'picture' => 'nullable|mimes:png,jpg,jpeg|max:5400',
             ]);
+            $password = $request->password;
+
             if($request->hasFile('picture')) {
                 $file = $request->file('picture');
                 $path = $file->store('images', 'public');
                 $validate['picture'] = $path;
             }
             $user = User::create($validate);
+            Mail::to($user->email)->send(new SendCredentials($user->id, $user->firstname));
 
-            if($request->permissions) {
-                $user->permissions()->sync($request->permissions);
-            }
             return response()->json([
                 'message' => 'User added successfully'
             ]);
@@ -160,15 +200,6 @@ class UserController extends Controller
             }
             $user->update($validate);
 
-            if($user->role && in_array($user->role->name, ['Unassigned', 'Rider', 'Delivery Rider'])) {
-                $user->permissions()->detach();
-            }
-
-            if($request->permissions) {
-                $user->permissions()->sync($request->permissions);
-            }else{
-                $user->permissions()->detach();
-            }
             return response()->json([
                 'message' => 'User updated successfully'
             ]);
@@ -206,9 +237,10 @@ class UserController extends Controller
     ///////////
 
     public function accountView() {
-        $user = User::with('role', 'permissions')->find(Auth::id());
+        $user = User::with('role')->find(Auth::id());
         return response()->json([
-            'user' => $user
+            'user' => $user,
+            'password' =>$user->password
         ]);
     }
 
@@ -223,12 +255,16 @@ class UserController extends Controller
                 'gender' => 'nullable',
                 'contact' => 'nullable',
                 'address' => 'nullable|string',
+                // 'password' => 'nullable',
                 'picture' => 'nullable|mimes:png,jpg,jpeg|max:5400',
             ]);
             if($request->hasFile('picture')) {
                 $file = $request->file('picture');
                 $path = $file->store('images', 'public');
                 $validate['picture'] = $path;
+            }
+            if($request->password) {
+                $validate['password'] = bcrypt($request->password);
             }
             $user->update($validate);
             return response()->json([
@@ -292,11 +328,38 @@ class UserController extends Controller
     //////////
 
     public function userMenuAccess() {
-        $user = User::with('permissions', 'role')->find(Auth::id());
+        $user = Auth::user();
+        $role = $user->role()->with('permissions')->first();
         $menu = Menu::with('permissions')->get();
         return response()->json([
-            'user' => $user,
+            'user' => $role,
             'menu' => $menu
+        ]);
+    }
+
+    public function dashboard() {
+        $user = Auth::user();
+        $role = $user->role()->with('permissions')->first();
+        $orders = Order::whereIn('status', ['Assigned', 'in_transit', 'to_be_deliver'])->count();
+        $transit = Order::where('status', 'in_transit')->count();
+        $myDelivery = Order::where('rider_id', Auth::user()->id)->whereIn('status', ['Assigned', 'in_transit'])->count();
+        $delivered = Order::where('status', 'Delivered')->count();
+        $users = User::whereHas('role', function($query) {
+            $query->where('name', '!=', 'Admin');
+        })->count();
+        $riders = User::whereHas('role', function($query) {
+            $query->whereIn('name', ['Rider', 'Food Rider', 'Delivery Rider', 'Food Delivery Rider']);
+        })->count();
+
+
+        return response()->json([
+            'role' => $role,
+            'users' => $users,
+            'orders' => $orders,
+            'riders' => $riders,
+            'transit' => $transit,
+            'myDelivery' => $myDelivery,
+            'delivered' => $delivered
         ]);
     }
 }
